+++ CHANGELOG.md
@@ -30,6 +30,10 @@
 
 ## Unreleased
 
+### Fixed
+
+- Crash when updating a song with 2 works existing with same title but different subtitle.
+
 ## 1.5.1 - 2019-12-06
 
 ### Fixed